# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Exam Project API",
    "description": "Documentation for the Exam Project API"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/admin/products": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Gets products page for admin.",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "type": "string"
          },
          {
            "name": "category",
            "in": "query",
            "type": "string"
          },
          {
            "name": "brand",
            "in": "query",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/admin/roles": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Gets roles page for admin.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/admin/orders": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Gets orders page for admin.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/orders/": {
      "get": {
        "tags": [
          "Orders"
        ],
        "description": "Gets all orders from the database for admins, and all orders a user has made for the user.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/orders/{orderId}": {
      "put": {
        "tags": [
          "Orders"
        ],
        "description": "Updates the status of a specific order.",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "description": "Deletes a specific order.",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "statusId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Gets all products from the database.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Creates a new product.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/products/{productId}": {
      "put": {
        "tags": [
          "Products"
        ],
        "description": "Updates a specific product.",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ProductEdit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "description": "Deletes a specific product.",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "isDeleted": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "imgUrl": {
                  "example": "any"
                },
                "unitPrice": {
                  "example": "any"
                },
                "quantity": {
                  "example": "any"
                },
                "categoryId": {
                  "example": "any"
                },
                "brandId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/roles/": {
      "get": {
        "tags": [
          "Roles"
        ],
        "description": "Gets all roles from the database.",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Roles"
        ],
        "description": "Creates a new role.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/roles/{roleId}": {
      "put": {
        "tags": [
          "Roles"
        ],
        "description": "Updates a specific role.",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Roles"
        ],
        "description": "Deletes a specific role.",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "role": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/statuses/": {
      "get": {
        "tags": [
          "Statuses"
        ],
        "description": "Gets all statuses from the database.",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Statuses"
        ],
        "description": "Creates a new status.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/statuses/{statusId}": {
      "put": {
        "tags": [
          "Statuses"
        ],
        "description": "Updates a specific status.",
        "parameters": [
          {
            "name": "statusId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Statuses"
        ],
        "description": "Deletes a specific status.",
        "parameters": [
          {
            "name": "statusId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Order": {
      "type": "object",
      "properties": {
        "statusId": {
          "type": "string",
          "example": "1"
        }
      },
      "required": [
        "statusId"
      ]
    },
    "Product": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Oculus Rift"
        },
        "description": {
          "type": "string",
          "example": "A VR Headset"
        },
        "imgUrl": {
          "type": "string",
          "example": "https://www.elkjop.no/image/dv_web_D180001002132666/OCULUSRIFT/oculus-rift-vr-headset.jpg"
        },
        "unitPrice": {
          "type": "string",
          "example": "699"
        },
        "quantity": {
          "type": "string",
          "example": "10"
        },
        "categoryId": {
          "type": "string",
          "example": "1"
        },
        "brandId": {
          "type": "string",
          "example": "1"
        }
      },
      "required": [
        "name",
        "description",
        "imgUrl",
        "unitPrice",
        "quantity",
        "categoryId",
        "brandId"
      ]
    },
    "ProductEdit": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Oculus Rift"
        },
        "description": {
          "type": "string",
          "example": "A VR Headset"
        },
        "imgUrl": {
          "type": "string",
          "example": "https://www.elkjop.no/image/dv_web_D180001002132666/OCULUSRIFT/oculus-rift-vr-headset.jpg"
        },
        "unitPrice": {
          "type": "string",
          "example": "699"
        },
        "quantity": {
          "type": "string",
          "example": "10"
        },
        "categoryId": {
          "type": "string",
          "example": "1"
        },
        "brandId": {
          "type": "string",
          "example": "1"
        },
        "isDeleted": {
          "type": "number",
          "example": 0
        }
      },
      "required": [
        "name",
        "description",
        "imgUrl",
        "unitPrice",
        "quantity",
        "categoryId",
        "brandId",
        "isDeleted"
      ]
    },
    "Role": {
      "type": "object",
      "properties": {
        "role": {
          "type": "string",
          "example": "Guest"
        }
      },
      "required": [
        "role"
      ]
    },
...
    "Status": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "On Hold"
        }
      },
      "required": [
        "status"
      ]
    },
...
  }
}